Rename InteractionService to IInteractor
