+++ OptiGUI/src/main/kotlin/opekope2/optigui/filter/factory/FilterFactory.kt
@@ -1,8 +1,8 @@
 package opekope2.optigui.filter.factory
 
 import net.fabricmc.fabric.api.event.player.UseBlockCallback
+import opekope2.optigui.interaction.IInteractor
 import opekope2.optigui.interaction.Interaction
-import opekope2.optigui.service.InteractionService
 import opekope2.optigui.util.identifier
 
 /**
@@ -17,7 +17,7 @@ import opekope2.optigui.util.identifier
  * most villager job sites), these preprocessors are not available. Processing either takes place in
  * [UseBlockCallback.interact], or in the filter created by [createFilter].
  * In the first case, check if the world is client-side, and if the targeted block is added by your mod
- * (check ID with [identifier]), call [InteractionService.interact].
+ * (check ID with [identifier]), call [IInteractor.interact].
  * In the latter case, you can use [Interaction.rawInteraction]. These details come from [UseBlockCallback.interact],
  * but this runs every game tick, instead of only once, when the interaction happens. You can pass a copy of the
  * interaction with modified details using [Interaction.copy].
