+++ OptiGUI/src/main/kotlin/opekope2/optigui/interaction/RawInteraction.kt
@@ -6,7 +6,6 @@ import net.minecraft.entity.player.PlayerEntity
 import net.minecraft.util.Hand
 import net.minecraft.util.hit.HitResult
 import net.minecraft.world.World
-import opekope2.optigui.service.InteractionService
 
 /**
  * Raw interaction details. When using preprocessors, this is likely unnecessary.
@@ -15,6 +14,6 @@ import opekope2.optigui.service.InteractionService
  *
  * @see UseBlockCallback.interact
  * @see UseEntityCallback.interact
- * @see InteractionService.interact
+ * @see IInteractor.interact
  */
 data class RawInteraction(val player: PlayerEntity, val world: World, val hand: Hand, val hitResult: HitResult?)
